+++ game-planetcoaster2/index.js
@@ -10,7 +10,6 @@ Date: 2025-10-22
 const { actions, fs, util, selectors, log } = require('vortex-api');
 const path = require('path');
 const template = require('string-template');
-//const Bluebird = require('bluebird');
 //const winapi = require('winapi-bindings'); //gives access to the Windows registry
 
 const USER_HOME = util.getVortexPath("home");